Exceptions does not wakeup queued requests
Async requests, when a connection is not readily available, they goes to
sleep. If the connection creation fails, the exception is not returned
to these requests.

Do not remember a failed request for a long time. If a connection
creation fails and if the exceptions are remembered for ever, new
requests will fail, regardless of whether they come in half a day
or half a month later. So drop the exceptions if not consumed for 5
seconds.

Wait in loop for connection

When you can create more connections, do not wait the entire time in a
queue as some other thread could have stolen the connection created. Now
the thread waits and tries to see if more connections can be created.

Also removed the logic for creating extra connections, as if there are
errors it causes too many premature failures. So a connection is created
only when requested.